+++ src/Helper/HelperUwPos.cpp
@@ -96,11 +96,6 @@ bool IsAtValeSpirits(const GW::GamePos &player_pos)
 
 bool IsInDhuumRoom(const GW::GamePos &player_pos)
 {
-#ifdef _DEBUG
-    if (GW::Map::GetMapID() == GW::Constants::MapID::Isle_of_the_Nameless)
-        return true;
-#endif
-
     return IsNearToGamePos(player_pos, GW::GamePos{-16105.50F, 17284.84F, 0}, GW::Constants::Range::Spellcast);
 }
